finished eventListeners, factory functions and findWinner function
This was a huge update. I have been working on this for many nights now trying to get all the code to function properly. I was getting tripped up and trying to rework code that wasn't my main issue. The main issue I was having in the initial tries of this project is that I was using the variable gameBoard to create all the divs in a for loop but was not actually changing any of the objects in the array but instead my event listeners were changing the innerHTML of the divs. After figuring that out, It took me a while to reorganize my code to work together again and on top of that trying to compile my code into factory functions. 

The project functionality is finally complete.